jay-backend/pyproject.toml
[project]
name = "jay-backend"
version = "0.1.0"
description = "JAY — AI-Powered Personal Skincare Companion API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pyjwt>=2.9.0",
    "cryptography>=44.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "psycopg2-binary>=2.9.11",
    "google-genai>=1.69.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
